RestAPI 적용하기
상품글 리뷰 컨트롤러 따로 빼서 정리하다 말았다
파라미터 이름 소문자만해서 어떻게 짓는지 찾아보기
페이지세팅이나, 리뷰갯수 세는 겻다리 서비스 로직 처리 어떻게 하는지도 알아보기
